pyproject.toml
[tool.poetry]
name = "flash_router"
version = "0.6.0-beta.2"
description = "File-system based router for Dash and Flash applications."
authors = ["chgiesse"]
repository = "https://github.com/chgiesse/dash-router"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.4"
dash = "^3.0.0rc1"
# Only Flash is optional - needed for flash_router package
dash-flash = "^1.0.0" 
pytest = "^9.0.2"

[tool.basedpyright]
typeCheckingMode = "off"
diagnosticMode = "workspace"
reportAny = "none"
reportArgumentType = "none"
reportAttributeAccessIssue = "none"
reportCallIssue = "none"
reportDeprecated = "none"
reportExplicitAny = "none"
reportGeneralTypeIssues = "none"
reportImplicitStringConcatenation = "none"
reportMissingTypeArgument = "none"
reportMissingTypeStubs = "none"
reportOptionalIterable = "none"
reportOperatorIssue = "none"
reportPrivateUsage = "none"
reportReturnType = "none"
reportUnannotatedClassAttribute = "none"
reportUnnecessaryIsInstance = "none"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"
reportUnusedCallResult = "none"
reportUnusedImport = "none"
reportUnusedVariable = "none"
reportUnreachable = "none"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
